Revert "More C casts to C++ casts."
This reverts commit 7524aa45f16b6579cc1ca207d64433924483f0ab.

+++ src/nupic/regions/TestNode.cpp
@@ -461,8 +461,7 @@ void TestNode::getParameterArray(const std::string &name, Int64 index, Array &ar
 	  if (index >= (Int64)unclonedInt64ArrayParam_.size()) {
       NTA_THROW << "uncloned parameter index out of range";
     }
-	  Array a(NTA_BasicType_Int64, &unclonedInt64ArrayParam_[static_cast<size_t>(index)][0], 
-                                  unclonedInt64ArrayParam_[static_cast<size_t>(index)].size());
+	  Array a(NTA_BasicType_Int64, &unclonedInt64ArrayParam_[(size_t)index][0], unclonedInt64ArrayParam_[(size_t)index].size());
     array = a;
   } else {
     NTA_THROW << "TestNode::getParameterArray -- unknown parameter " << name;
@@ -486,10 +485,10 @@ void TestNode::setParameterUInt32(const std::string &name, Int64 index, UInt32 v
     uint32Param_ = value;
   }
   else if (name == "unclonedParam") {
-    if (index < 0 || index >= static_cast<Int64>(unclonedParam_.size())) {
+    if (index < 0 || index >= (Int64)unclonedParam_.size()) {
       NTA_THROW << "uncloned parameters index out of range";
     }
-	unclonedParam_[static_cast<size_t>(index)] = value;
+	unclonedParam_[(size_t)index] = value;
   }
   else if (name == "shouldCloneParam") {
       shouldCloneParam_ = !(value == 0);
@@ -498,7 +497,7 @@ void TestNode::setParameterUInt32(const std::string &name, Int64 index, UInt32 v
     if (shouldCloneParam_) {
       possiblyUnclonedParam_[0] = value;
     } else {
-      if (index < 0 || index >= static_cast<Int64>(possiblyUnclonedParam_.size())) {
+      if (index < 0 || index >= (Int64)possiblyUnclonedParam_.size()) {
         NTA_THROW << "uncloned parameters index out of range.";
       }
       possiblyUnclonedParam_[(size_t)index] = value;
@@ -576,7 +575,7 @@ void TestNode::setParameterArray(const std::string &name, Int64 index, const Arr
     if (index < 0) {
       NTA_THROW << "uncloned parameters cannot be accessed at region level";
     }
-    unclonedInt64ArrayParam_[static_cast<Int64>(index)] = array.asVector<Int64>();
+    unclonedInt64ArrayParam_[(size_t)index] = array.asVector<Int64>();
   }
   else {
     RegionImpl::setParameterArray(name, index, array);
@@ -595,7 +594,7 @@ size_t TestNode::getParameterArrayCount(const std::string &name, Int64 index) {
     if (index < 0) {
       NTA_THROW << "uncloned parameters cannot be accessed at region level";
     }
-    return unclonedInt64ArrayParam_[index].size();
+    return unclonedInt64ArrayParam_[(UInt)index].size();
   } else {
     NTA_THROW << "TestNode::getParameterArrayCount -- unknown parameter "
               << name;
